# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  models.EvaluasiDosen:
    properties:
      _id:
        type: integer
      Inovasi_dalam_perkuliahan:
        $ref: '#/definitions/models.Penilaian'
      Kedisiplinan:
        $ref: '#/definitions/models.Penilaian'
      Kemampuan_berinteraksi_sosial_dengan_mahasiswa:
        $ref: '#/definitions/models.Penilaian'
      Kemampuan_berkomunikasi:
        $ref: '#/definitions/models.Penilaian'
      Keteladanan_dalam_bersikap_dan_bertindak:
        $ref: '#/definitions/models.Penilaian'
      Kreatifitas_dalam_perkuliahan:
        $ref: '#/definitions/models.Penilaian'
      Media_pembelajaran:
        $ref: '#/definitions/models.Penilaian'
      Metode_dan_pendekatan_perkuliahan:
        $ref: '#/definitions/models.Penilaian'
      Motivasi_dan_antusiasme_mengajar:
        $ref: '#/definitions/models.Penilaian'
      Pemberian_tugas_perkuliahan:
        $ref: '#/definitions/models.Penilaian'
      Penciptaan_iklim_belajar:
        $ref: '#/definitions/models.Penilaian'
      Penegakan_aturan_perkuliahan:
        $ref: '#/definitions/models.Penilaian'
      Pengelolaan_kelas:
        $ref: '#/definitions/models.Penilaian'
      Pengembangan_karakter_mahasiswa:
        $ref: '#/definitions/models.Penilaian'
      Penggunaan_bahasa_lisan_dan_tulisan:
        $ref: '#/definitions/models.Penilaian'
      Penguasaan_materi_perkuliahan:
        $ref: '#/definitions/models.Penilaian'
      Penilaian_hasil_belajar:
        $ref: '#/definitions/models.Penilaian'
      Penilaian_proses_belajar:
        $ref: '#/definitions/models.Penilaian'
      Perencanaan_perkuliahan:
        $ref: '#/definitions/models.Penilaian'
      Relevansi_materi_dengan_tujuan_pembelajaran:
        $ref: '#/definitions/models.Penilaian'
      Sumber_belajar:
        $ref: '#/definitions/models.Penilaian'
      created_at:
        type: string
      deleted_at:
        type: string
      kode_fakultas:
        type: string
      kode_matakuliah:
        type: string
      kode_prodi:
        type: string
      nama_kelas:
        type: string
      nama_lengkap:
        type: string
      nama_matakuliah:
        type: string
      nip:
        type: string
      no_induk_undiksha:
        type: string
      semester:
        type: string
      tahun:
        type: string
      unit:
        $ref: '#/definitions/models.Unit'
      updated_at:
        type: string
    type: object
...
  models.Hki:
    properties:
      _id:
        type: integer
      anggota_penelitian:
        type: string
      create_dosen_id:
        type: string
      created_at:
        type: string
      cron_semester:
        type: string
      cron_tahun:
        type: string
      deleted_at:
        type: string
      deskripsi:
        type: string
      file_bukti_kinerja:
        type: string
      file_hasil_uji_plagiarim:
        type: string
      file_pemeriksaan_substansi:
        type: string
      file_pendaftaran:
        type: string
      file_penilaian_reviewer:
        type: string
      file_sertifikasi:
        type: string
      file_sertifikat_paten:
        type: string
      file_uji_publik:
        type: string
      is_produk:
        type: string
      isValid:
        type: string
      jenis_paten:
        type: string
      jml_negara_pengaku:
        type: string
      jml_penulis:
        type: string
      kode_jenis_paten:
        type: string
      kode_scope:
        type: string
      komentar:
        type: string
      level_capaian:
        type: string
      mahasiswa_penelitian:
        type: string
      nama_dosen:
        type: string
      nama_karya:
        type: string
      no_pendaftaran:
        type: string
      no_pendatatan_sertifikat:
        type: string
      periode:
        type: string
      posisi:
        type: string
      primaryKey:
        type: string
      produk_penelitian:
        type: string
      produk_penelitian_id:
        type: string
      produk_penelitian_judul:
        type: string
      produk_pengabdian:
        type: string
      produk_pengabdian_id:
        type: string
      produk_pengabdian_judul:
        type: string
      scope:
        type: string
      semester:
        type: string
      semester_type:
        type: string
      sumber_produk:
        type: string
      tahun_ajaran:
        type: string
      tahun_data:
        type: string
      tanggal:
        type: string
      tbName:
        type: string
      units:
        $ref: '#/definitions/models.Unit'
      updated_at:
        type: string
      valid_ipk:
        type: string
      valid_ipk_komentar:
        type: string
      waktu_pelaksanaan:
        type: string
    type: object
...
  models.ListResponse:
    properties:
      datas: {}
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
      status:
        example: success
        type: string
    type: object
...
  models.PaginationResponse:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      pages:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  models.Penilaian:
    properties:
      baik:
        type: integer
      cukup:
        type: integer
      kurang:
        type: integer
      sangat_baik:
        type: integer
    type: object
...
  models.Unit:
    properties:
      fakultas:
        type: string
      fkt_kode:
        type: string
      jrs_kode:
        type: string
      jurusan:
        type: string
      prd_kode:
        type: string
      prodi:
        type: string
      uk_kode:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: anggadek857@gmail.com
    name: Dwi Angga
    url: https://github.com/anggananda
  description: |-
    This repository contains the **RESTful API** for the **Executive Information System (EIS) Undiksha**, built with **Golang** using the **Gin framework**.
    The API consumes data from **MongoDB** (populated via ETL from SQL in a separate repository) and uses **Redis** for caching.
    The system is integrated with **Undiksha Single Sign-On (SSO)** and implements **JWT Authentication** for secure access.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/anggananda
  title: Executive Information System RESTful API
  version: "1.0"
paths:
...
  /evaluasi-dosen:
    get:
      consumes:
      - application/json
      description: Mendapatkan data Evaluasi Dosen berdasarkan filter dengan pagination
      parameters:
      - default: 1
        description: 'Nomor halaman (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Jumlah data per halaman (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter berdasarkan kode fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan kode jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan kode program studi
        in: query
        name: kodeProdi
        type: string
      - description: Filter berdasarkan tahun akademik
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: string
      - description: Filter berdasarkan namaDosen
        in: query
        name: namaDosen
        type: string
      - description: Pencarian global
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListResponse'
            - properties:
                datas:
                  items:
                    $ref: '#/definitions/models.EvaluasiDosen'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get filtered Evaluasi Dosen
      tags:
      - Evaluasi Dosen
  /evaluasi-dosen/export-csv:
    get:
      consumes:
      - application/json
      description: Mengekspor daftar Evaluasi dosen yang telah difilter ke dalam file
        CSV.
      parameters:
      - description: Maksimal data yang akan diekspor
        in: query
        name: limit
        type: integer
      - description: Filter berdasarkan Kode Fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan Kode Jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan Kode Program Studi
        in: query
        name: kodeProdi
        type: string
      - description: 'Filter berdasarkan Tahun Ajaran (default: tahun sekarang)'
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: string
      - description: Filter berdasarkan namaDosen
        in: query
        name: namaDosen
        type: string
      - description: Pencarian bebas
        in: query
        name: search
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File CSV berhasil diunduh
          schema:
            type: file
        "500":
          description: Kesalahan pada server saat pengambilan data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export evaluasi Dosen ke CSV
      tags:
      - Evaluasi Dosen
...
  /hki:
    get:
      consumes:
      - application/json
      description: Mendapatkan data hki berdasarkan filter dengan pagination
      parameters:
      - default: 1
        description: 'Nomor halaman (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Jumlah data per halaman (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Filter berdasarkan kode fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan kode jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan kode program studi
        in: query
        name: kodeProdi
        type: string
      - description: Filter berdasarkan tahun akademik
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: string
      - description: Pencarian global
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ListResponse'
            - properties:
                datas:
                  items:
                    $ref: '#/definitions/models.Hki'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get filtered hki
      tags:
      - Hki
  /hki/export-csv:
    get:
      consumes:
      - application/json
      description: Mengekspor daftar hki yang telah difilter ke dalam file CSV.
      parameters:
      - description: Maksimal data yang akan diekspor
        in: query
        name: limit
        type: integer
      - description: Filter berdasarkan Kode Fakultas
        in: query
        name: kodeFakultas
        type: string
      - description: Filter berdasarkan Kode Jurusan
        in: query
        name: kodeJurusan
        type: string
      - description: Filter berdasarkan Kode Program Studi
        in: query
        name: kodeProdi
        type: string
      - description: 'Filter berdasarkan Tahun Ajaran (default: tahun sekarang)'
        in: query
        name: tahun
        type: string
      - description: Filter berdasarkan semester
        in: query
        name: semester
        type: string
      - description: Pencarian bebas
        in: query
        name: search
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File CSV berhasil diunduh
          schema:
            type: file
        "500":
          description: Kesalahan pada server saat pengambilan data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Export Hki ke CSV
      tags:
      - Hki
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"